+++ bindings/encoding/struct_test.go
@@ -2,6 +2,7 @@ package encoding
 
 import (
 	"crypto/rand"
+	"github.com/taikoxyz/taiko-client/internal/utils"
 	"math/big"
 	"testing"
 	"time"
@@ -10,8 +11,6 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/stretchr/testify/require"
-
-	"github.com/taikoxyz/taiko-client/common/utils"
 )
 
 var (
