+++ driver/state/state_test.go
@@ -2,6 +2,7 @@ package state
 
 import (
 	"context"
+	"github.com/taikoxyz/taiko-client/internal/utils"
 	"math/big"
 	"testing"
 
@@ -9,7 +10,6 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/taikoxyz/taiko-client/common/utils"
 	"github.com/taikoxyz/taiko-client/internal/testutils"
 )
