+++ internal/utils/utils.go
@@ -29,5 +29,5 @@ func RandUint32(max *big.Int) uint32 {
 
 // IsNil checks if the interface is empty.
 func IsNil(i interface{}) bool {
-	return i == nil || reflect2.IsNil(i)
+	return reflect2.IsNil(i)
 }
